requirements.txt
python>=3.6
torch>=1.7.0
torchvision>=0.8.1
jittor==1.3.4.12
scipy==1.7.3
tensorboardX==2.5.1
opencv_python==4.5.5.64
Pillow==9.2.0
tqdm==4.64.0
numpy